requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic-settings==2.5.2
alembic==1.13.2
python-multipart==0.0.9
